+++ toolsrc/include/vcpkg_Environment.h
@@ -14,4 +14,14 @@ namespace vcpkg::Environment
         ensure_cmake_on_path(paths);
         ensure_git_on_path(paths);
     }
+
+    const fs::path& get_dumpbin_exe(const vcpkg_paths& paths);
+
+    struct vcvarsall_and_platform_toolset
+    {
+        fs::path path;
+        std::wstring platform_toolset;
+    };
+
+    const vcvarsall_and_platform_toolset& get_vcvarsall_bat(const vcpkg_paths& paths);
 }